requirements.txt
streamlit 
langchain 
langchain_google_genai
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv